+++ kitty/config_data.py
@@ -656,6 +656,14 @@ probably better off just hiding the titlebar with :opt:`macos_hide_titlebar`.
 o('macos_hide_titlebar', False, long_text=_('''
 Hide the kitty window's title bar on macOS.'''))
 
+
+o('x11_hide_window_decorations', False, long_text=_('''
+Hide the window decorations (title bar and window borders) on X11 and Wayland.
+Whether this works and exactly what effect it has depends on the window
+manager, as it is the job of the window manager/compositor to draw window
+decorations.'''))
+
+
 o('macos_option_as_alt', True, long_text=_('''
 Use the option key as an alt key. With this set to no, kitty will use
 the macOS native :kbd:`Option+Key` = unicode character behavior. This will